refactor: Move FDv2 compound types into a model module
