+++ launchdarkly-server-sdk/src/fdv2/mod.rs
@@ -1,2 +1,3 @@
+mod model;
 mod protocol;
 mod wire;
