+++ launchdarkly-server-sdk/src/fdv2/wire.rs
@@ -57,32 +57,6 @@ pub(super) struct FDv2Error {
     pub(super) reason: String,
 }
 
-#[derive(Debug, PartialEq, Eq)]
-pub(super) enum Selector {
-    Empty,
-    Set { state: String },
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) enum ChangeSetKind {
-    None,
-    Full,
-    Partial,
-}
-
-#[derive(Debug)]
-pub(super) enum FDv2Change {
-    Put(PutObject),
-    Delete(DeleteObject),
-}
-
-#[derive(Debug)]
-pub(super) struct ChangeSet {
-    pub(super) kind: ChangeSetKind,
-    pub(super) changes: Vec<FDv2Change>,
-    pub(super) selector: Selector,
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
